+++ src/RestaurantContainer.jsx
@@ -4,47 +4,14 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import RestaurantDetail from './RestaurantDetail';
 
+import ReviewForm from './ReviewForm';
+
 import {
   loadRestaurant,
   changeReviewField,
 } from './actions';
 
 import { get } from './utils';
-
-function ReviewForm({ onChange }) {
-  function handleChange(event) {
-    const { target: { name, value } } = event;
-
-    onChange({ name, value });
-  }
-
-  return (
-    <>
-      <div>
-        <label htmlFor="reivew-score">
-          평점
-        </label>
-        <input
-          type="number"
-          id="reivew-score"
-          name="score"
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <label htmlFor="reivew-description">
-          리뷰 내용
-        </label>
-        <input
-          type="text"
-          id="reivew-description"
-          name="description"
-          onChange={handleChange}
-        />
-      </div>
-    </>
-  );
-}
 
 export default function RestaurantContainer({ restaurantId }) {
   const dispatch = useDispatch();
